+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Performance
+
+- Avoid allocations in error formatting.
+
 ## [0.7.2] - 2021-06-22
 
 ### Fixed